컴포넌츠 분할 완료
