+++ public/data/Data.json
@@ -1,0 +1,45 @@
+{
+  "feed": [
+    {
+      "id": 1,
+      "feedUpLoader": "Xabi",
+      "uploaderProfileImg": "images/yoonkyeong/uploaderImage/uploader1.jpg",
+      "feedImage": "images/yoonkyeong/feed/chorong-crop.jpg",
+      "comments": [
+        {
+          "id": 1,
+          "user": "Mathias",
+          "comment": "fehlst mir!"
+        },
+        {
+          "id": 2,
+          "user": "Yue",
+          "comment": "long time no see!"
+        },
+        {
+          "id": 3,
+          "user": "Ruonan",
+          "comment": "wwww"
+        }
+      ]
+    },
+    {
+      "id": 2,
+      "feedUpLoader": "ZZabi",
+      "uploaderProfileImg": "images/yoonkyeong/uploaderImage/uploader2.jpg",
+      "feedImage": "images/yoonkyeong/feed/xabi1-crop.jpg",
+      "comments": [
+        {
+          "id": 1,
+          "user": "lilly",
+          "comment": "schon tot"
+        },
+        {
+          "id": 2,
+          "user": "klar",
+          "comment": "집가고싶어!"
+        }
+      ]
+    }
+  ]
+}
