--- public/data/commentData.json
@@ -1,20 +1,0 @@
-[
-  {
-    "id": 1,
-    "userName": "wecode",
-    "content": "Welcome to world best coding bootcamp!",
-    "isLiked": true
-  },
-  {
-    "id": 2,
-    "userName": "joonsikyang",
-    "content": "Hi there.",
-    "isLiked": false
-  },
-  {
-    "id": 3,
-    "userName": "jayPark",
-    "content": "Hey.",
-    "isLiked": false
-  }
-]
