+++ src/pages/yoonkyeong/Main/Feeds/Feed/Article/Article.scss
@@ -1,0 +1,72 @@
+@use '../../../../../../styles/mixin' as mixin;
+@use '../../../../../../styles/color' as color;
+
+.article {
+  background-color: #fff;
+
+  .articleHeader {
+    display: flex;
+    justify-content: space-between;
+    padding: 14px 4px 14px 16px;
+
+    .publisherProfile {
+      display: flex;
+
+      .publisherImage {
+        width: 32px;
+        height: 32px;
+        border-radius: 50%;
+        overflow: hidden;
+
+        img {
+          width: 100%;
+        }
+      }
+
+      .publisherId {
+        text-align: center;
+        padding: 5px 15px;
+        font-weight: bold;
+      }
+    }
+  }
+
+  .articleWrapper {
+    width: 615px;
+    height: 615px;
+
+    .articleImage {
+      width: 100%;
+      height: 100%;
+    }
+  }
+
+  .articleIconsWrapper {
+    display: flex;
+    justify-content: space-around;
+    position: relative;
+    width: 140px;
+    padding: 10px 0px;
+
+    img {
+      height: 100%;
+      padding: 10px;
+    }
+
+    .bookmark {
+      position: absolute;
+      right: -470px;
+    }
+  }
+
+  .likeWrapper {
+    padding: 0px 16px;
+    margin: 5px 0px 15px 0px;
+  }
+  .timeWrapper {
+    padding-left: 16px;
+    margin: 5px 0px 16px 0px;
+    font-size: smaller;
+    color: #b4b4b4;
+  }
+}
