+++ src/pages/yoonkyeong/Main/Feeds/Feed/CommentWrapper/Comment/comment.scss
@@ -1,5 +1,5 @@
-@use '../../../styles/mixin' as mixin;
-@use '../../../styles/color' as color;
+@use '../../../../../../../styles/mixin' as mixin;
+@use '../../../../../../../styles/color' as color;
 
 li {
   display: flex;
