+++ src/pages/yoonkyeong/Main/Feeds/Feed/Feed.scss
@@ -1,0 +1,9 @@
+@use '../../../../../styles/mixin' as mixin;
+@use '../../../../../styles/color' as color;
+
+.feed {
+  display: flex;
+  flex-flow: column wrap;
+  border: 1px solid color.$clr-border;
+  margin-bottom: 30px;
+}
