--- src/pages/yoonkyeong/component/CommentList.scss
@@ -1,3 +1,0 @@
-.mockData {
-  color: lightgrey;
-}
